Initial parallelization of TVDLimit
Compiles and runs on Titan x. There's a handful of branching logic (for debug specifically) that could be handled either with an ifdef at compile time or an alternate function at runtime. Leaving it alone for now. There was one change to one of the loops that could cause an error in the results: this loop is bounded by a variable called nvar, which seems like it will always be 8. I need to double check, but it has not caused problems thus far.